Update README.md instructions clear
The instructions are clear enough now and some error handling
has been added, especially the build is now correct (I think).
