+++ .env.example
@@ -1,2 +1,3 @@
 NEXT_PUBLIC_USER_1=User1
 NEXT_PUBLIC_USER_2=User2
+DATA_USERS_PATH=./data/users.json
